requirements/dev.txt
-r base.txt
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
debugpy>=1.8,<2.0
